# src/main/openapi/api.yaml
openapi: 3.0.3
info:
  title: Application lifecycle API
  version: 0.0.1
  description: >-
    Strict application lifecycle: CREATED to VERIFIED to ACCEPTED to PUBLISHED;
    CREATED to DELETED; VERIFIED or ACCEPTED to REJECTED. All other transitions,
    including repeated actions, return 409. Body edits are allowed only in CREATED
    or VERIFIED. PUBLISHED, DELETED, and REJECTED are terminal states. Publication
    assigns a unique positive publicationNumber; numbering may contain gaps.
tags:
  - name: Applications
paths:
  /applications:
    get:
      tags: [Applications]
      operationId: listApplications
      summary: List applications with pagination and optional filters
      description: >-
        Filters combine with AND. Name matching is a case-insensitive literal substring
        search; omitted or blank names impose no name restriction. State matching is
        exact and case-sensitive. DELETED applications are excluded unless state=DELETED
        is explicitly requested. Results are ordered by createdAt descending, then UUID
        descending. Empty results and pages beyond the results return 200 with empty content.
      parameters:
        - name: name
          in: query
          description: Case-insensitive literal substring of the application name
          schema:
            type: string
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/ApplicationState'
        - name: page
          in: query
          description: Zero-based page index
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Number of applications per page
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Page of matching applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Applications]
      operationId: createApplication
      summary: Create an application
      description: Assigns a UUID and the initial CREATED state. Clients cannot supply a state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /applications/{id}:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    patch:
      tags: [Applications]
      operationId: editApplication
      summary: Replace the application body without changing its name
      description: >-
        Allowed only in CREATED or VERIFIED, without changing the state or name.
        Editing an ACCEPTED, PUBLISHED, or REJECTED application returns 409;
        editing a DELETED application returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditApplicationRequest'
      responses:
        '200':
          description: Application updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Applications]
      operationId: deleteApplication
      summary: Soft-delete an application with a mandatory reason
      description: >-
        Allowed only in CREATED. Sets the state to DELETED and retains the application,
        deletion timestamp, and reason. Repeated deletion returns 409 without changing data.
      parameters:
        - name: reason
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/DeletionReason'
      responses:
        '204':
          description: Application soft-deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /applications/{id}/verification:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    put:
      tags: [Applications]
      operationId: verifyApplication
      summary: Verify a created application
      description: Only CREATED to VERIFIED is allowed. Repeated verification returns 409.
      responses:
        '200':
          $ref: '#/components/responses/ApplicationChanged'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /applications/{id}/acceptance:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    put:
      tags: [Applications]
      operationId: acceptApplication
      summary: Accept a verified application
      description: Only VERIFIED to ACCEPTED is allowed. Repeated acceptance returns 409.
      responses:
        '200':
          $ref: '#/components/responses/ApplicationChanged'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /applications/{id}/publication:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    put:
      tags: [Applications]
      operationId: publishApplication
      summary: Publish an accepted application
      description: >-
        Only ACCEPTED to PUBLISHED is allowed. Atomically assigns a unique positive
        publicationNumber from a database sequence and enters the terminal PUBLISHED
        state. Numbers may contain gaps. UUIDs remain the identifiers used in URLs.
        Repeated publication returns 409 without changing or allocating another number.
      responses:
        '200':
          $ref: '#/components/responses/ApplicationChanged'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /applications/{id}/rejection:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    put:
      tags: [Applications]
      operationId: rejectApplication
      summary: Reject a verified or accepted application
      description: >-
        Allowed only in VERIFIED or ACCEPTED. A nonblank reason is mandatory.
        Sets the state to REJECTED. Repeated rejection returns 409; its original
        reason and timestamp cannot be replaced.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectApplicationRequest'
      responses:
        '200':
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  parameters:
    ApplicationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ApplicationChanged:
      description: Application state changed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApplicationResponse'
    BadRequest:
      description: Missing or invalid input, including unexpected JSON properties
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    NotFound:
      description: Application does not exist, or is deleted when editing its body
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    Conflict:
      description: >-
        Application has already been changed or the action is not allowed in its
        current state. No data, reasons, or timestamps are modified.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
  schemas:
    ApplicationPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponse'
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int64
          minimum: 0
    ApplicationState:
      type: string
      enum: [CREATED, VERIFIED, ACCEPTED, PUBLISHED, DELETED, REJECTED]
    CreateApplicationRequest:
      type: object
      additionalProperties: false
      required: [name, body]
      properties:
        name:
          type: string
          minLength: 1
          pattern: '[\s\S]*\S[\s\S]*'
          example: Loan application
        body:
          type: string
          minLength: 1
          pattern: '[\s\S]*\S[\s\S]*'
          example: Please review my application.
    EditApplicationRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          pattern: '[\s\S]*\S[\s\S]*'
    RejectApplicationRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          pattern: '[\s\S]*\S[\s\S]*'
          example: Required documents are missing.
    DeletionReason:
      type: string
      enum: [DUPLICATE, CREATED_BY_MISTAKE, NO_LONGER_NEEDED]
    ApplicationResponse:
      type: object
      required: [id, name, body, state, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        state:
          $ref: '#/components/schemas/ApplicationState'
        publicationNumber:
          type: integer
          format: int64
          minimum: 1
          nullable: true
          readOnly: true
          description: >-
            Unique numeric identifier assigned only on successful publication;
            null in all other states. Cannot be provided or modified by clients.
            Numbering may contain gaps.
          example: 123
        updatedAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
          nullable: true
        rejectedAt:
          type: string
          format: date-time
          nullable: true
    ApiProblem:
      type: object
      description: RFC 9457 problem details
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri